samplepy02/pyproject.toml
[project]
name = "samplepy02"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
    "selenium>=4.35.0",
]
